UpdateRoom ErrorFixed
password is now limited to 4
update room with empty hashtags work

+++ src/room/room.service.ts
@@ -142,6 +142,7 @@ export class RoomService {
   async updateRoom(file, roomId, updateRoomDto) {
     const targetRoom = await this.findRoom(roomId);
     let filename = targetRoom.imageLocation;
+    let hashtags = targetRoom.hashtags;
 
     //if new image is provided, delete original image
     if (file) {
@@ -160,42 +161,44 @@ export class RoomService {
     }
 
     //reset Hashtags
+    if (updateRoomDto.hashtag) {
+      for (let i = 0; i < targetRoom.hashtags.length; i++) {
+        const tag = targetRoom.hashtags[i];
+        const dbHashtag = await this.findHashtag(tag);
 
-    for (let i = 0; i < targetRoom.hashtags.length; i++) {
-      const tag = targetRoom.hashtags[i];
-      const dbHashtag = await this.findHashtag(tag);
+        //If hashtag length is 1, delete hashtag from DB, else, remove roomId from hashtag.rooms
 
-      //If hashtag length is 1, delete hashtag from DB, else, remove roomId from hashtag.rooms
-
-      if (dbHashtag.rooms.length == 1) {
-        await this.hashtagModel.deleteOne({ _id: dbHashtag._id }).exec();
-      } else {
-        await this.hashtagModel.updateOne(
-          { _id: dbHashtag._id },
-          { $pull: { rooms: roomId } },
-        );
+        if (dbHashtag.rooms.length == 1) {
+          await this.hashtagModel.deleteOne({ _id: dbHashtag._id }).exec();
+        } else {
+          await this.hashtagModel.updateOne(
+            { _id: dbHashtag._id },
+            { $pull: { rooms: roomId } },
+          );
+        }
       }
-    }
 
-    //change JSON Object to Array Object
-    const hashtags = updateRoomDto.hashtag
-      .toString()
-      .replace(/\[|\]/g, '')
-      .replace(/\s/g, '')
-      .split(',');
+      //change JSON Object to Array Object
 
-    //recreate Hashtags
-    for (let i = 0; i < hashtags.length; i++) {
-      const tag = hashtags[i];
-      const dbHashtag = await this.findHashtag(tag);
+      hashtags = updateRoomDto.hashtag
+        .toString()
+        .replace(/\[|\]/g, '')
+        .replace(/\s/g, '')
+        .split(',');
 
-      //if no hashtag exists, create one
-      if (!dbHashtag) {
-        const newHashtag = new this.hashtagModel({
-          content: tag,
-          rooms: [roomId as string],
-        });
-        await newHashtag.save();
+      //recreate Hashtags
+      for (let i = 0; i < hashtags.length; i++) {
+        const tag = hashtags[i];
+        const dbHashtag = await this.findHashtag(tag);
+
+        //if no hashtag exists, create one
+        if (!dbHashtag) {
+          const newHashtag = new this.hashtagModel({
+            content: tag,
+            rooms: [roomId as string],
+          });
+          await newHashtag.save();
+        }
       }
     }
 
@@ -231,7 +234,6 @@ export class RoomService {
     let hashtag;
     try {
       hashtag = await this.hashtagModel.findOne({ content }).exec();
-      console.log('wassup', hashtag);
     } catch (error) {
       return null;
     }
